add panels on mainform

+++ ninetyFourPercent/Data/PlayerInfo.cs
@@ -2,8 +2,8 @@
 {
     public static class PlayerInfo
     {
-        public static int ID { get; set; }
+        public static long ID { get; set; }
         public static string LOGIN { get; set; }
-        public static int MONEY { get; set; }
+        public static long MONEY { get; set; }
     }
 }
